WEBAPP-375: Remove date and location information to PageDetail
