+++ src/modules/common/components/Page.js
@@ -1,26 +1,53 @@
 // @flow
 
-import React from 'react'
-import Caption from './Caption'
+import * as React from 'react'
+import type { TFunction } from 'react-i18next'
+import { translate } from 'react-i18next'
+import styled from 'styled-components'
 import RemoteContent from './RemoteContent'
+import Caption from './Caption'
+import type Moment from 'moment'
+import LastUpdateInfo from './LastUpdateInfo'
+
+const Thumbnail = styled.img`
+  display: flex;
+  width: 300px;
+  height: 300px;
+  margin: 10px auto;
+  padding-bottom: 10px;
+  object-fit: contain;
+`
 
 type PropsType = {|
   title: string,
+  thumbnail?: string,
   content: string,
-  onInternLinkClick: string => void,
-  hijackRegExp?: RegExp
+  lastUpdate: Moment,
+  language: string,
+  t: TFunction,
+  onInternalLinkClick: string => void,
+  hijackRegExp?: RegExp,
+  children?: Array<React.Node>
 |}
 
-class Page extends React.Component<PropsType> {
+/**
+ * Display a single page with all necessary information
+ */
+export class Page extends React.Component<PropsType> {
   render () {
+    const {title, thumbnail, content, lastUpdate, language, hijackRegExp, children, onInternalLinkClick} = this.props
     return (
-      <div>
-        <Caption title={this.props.title} />
-        <RemoteContent dangerouslySetInnerHTML={{__html: this.props.content}} centered={false}
-                       onInternLinkClick={this.props.onInternLinkClick} hijackRegExp={this.props.hijackRegExp} />
-      </div>
+      <>
+        {thumbnail && <Thumbnail src={thumbnail} />}
+        <Caption title={title} />
+        {children}
+        <RemoteContent dangerouslySetInnerHTML={{__html: content}}
+                       onInternLinkClick={onInternalLinkClick}
+                       hijackRegExp={hijackRegExp} />
+        <LastUpdateInfo lastUpdate={lastUpdate} language={language} />
+      </>
     )
   }
 }
 
-export default Page
+export default translate('common')(Page)
