+++ src/modules/common/components/__tests__/Page.spec.js
@@ -2,11 +2,27 @@
 
 import React from 'react'
 import { shallow } from 'enzyme'
-import Page from '../Page'
+import moment from 'moment-timezone'
+import { Page } from '../Page'
 
 describe('Page', () => {
+  const title = 'first Event'
+  const lastUpdate = moment('2016-01-07 10:36:24')
+  const content = 'content'
+  const thumbnail = 'thumbnail'
+
+  const language = 'en'
+
+  const t = (key: ?string): string => key || ''
+
   it('should render', () => {
-    expect(shallow(<Page title={'Test Title'} content={'This is a test content'}
-                         onInternLinkClick={() => {}} />)).toMatchSnapshot()
+    expect(shallow(<Page title={title}
+                         lastUpdate={lastUpdate}
+                         content={content}
+                         thumbnail={thumbnail}
+                         language={language}
+                         onInternalLinkClick={() => {}}
+                         t={t} />
+    )).toMatchSnapshot()
   })
 })
